requirements.txt
streamlit
pdfplumber
pandas
xlsxwriter
docx2txt
pytesseract
Pillow